requirements.txt
asgiref==3.4.1
attrs==21.2.0
Automat==20.2.0
certifi==2021.5.30
charset-normalizer==2.0.3
constantly==15.1.0
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.7.0
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.3.0
zope.interface==5.4.0
